Large refactor including a graphics api and context abstraction to the window main has now been simplified and intialization and main loop lives within Application

+++ Chroma/Chroma/source/Engine/animation/Animator.cpp
@@ -53,7 +53,7 @@ float Animator::CalculateFrameNumber(std::string const& takeName, float const& n
 	// TODO: update framenumber interpolation
 	if (normalizedTime == 1.0)
 	{
-		return Chroma::Math::Remap(0.999, 0.0, 1.0, 0.0, (float)m_Takes.at(takeName).m_NumFrames);
+		return Chroma::Math::Remap(0.999, 0.0, 1.0, 0.0, (double)m_Takes.at(takeName).m_NumFrames);
 	}
 	else
 		return Chroma::Math::Remap(normalizedTime, 0.0, 1.0, 0.0, (float)m_Takes.at(takeName).m_NumFrames);
